+++ frontend/src/components/trigger/Actions.test.tsx
@@ -37,6 +37,7 @@ describe("ActionForm", () => {
       isWritable: true,
       isMandatory: false,
       isDeleteInChain: false,
+      isSummarized: false,
       referral: [],
     })),
     webhooks: [],